Update LICENSE
This commit updates the project's license from the MIT License to the GNU General Public License version 3 (GPL-3.0). The license text has been replaced with the GPL-3.0 license text to reflect the change in licensing terms. This update ensures that the project complies with the GPL-3.0 license requirements and is aligned with the project's licensing goals.